fixed package details missing, removed extra spacing on admin package change, used same button component across app and restyled buttons slightly
