+++ src/services/ui/src/components/Form/fields/index.ts
@@ -1,3 +1,2 @@
 export * from "./AdditionalInfoSection";
 export * from "./AttachmentsSection";
-export * from "./FormButtons";
