+++ src/services/ui/src/features/submission/waiver/capitated/capitated-1915-b-waiver-amendment.tsx
@@ -273,7 +273,11 @@ export const Capitated1915BWaiverAmendmentPage = () => {
             />
           </SectionCard>
           <Content.PreSubmissionMessage />
-          <SubmitAndCancelBtnSection />
+          <SubmitAndCancelBtnSection
+            showAlert
+            loadingSpinner
+            cancelNavigationLocation={originPath ?? "/dashboard"}
+          />
         </form>
       </Inputs.Form>
       <FAQFooter />
