+++ src/services/ui/src/features/submission/waiver/capitated/capitated-1915-b-waiver-initial.tsx
@@ -243,7 +243,11 @@ export const Capitated1915BWaiverInitialPage = () => {
             />
           </SectionCard>
           <Content.PreSubmissionMessage />
-          <SubmitAndCancelBtnSection />
+          <SubmitAndCancelBtnSection
+            showAlert
+            loadingSpinner
+            cancelNavigationLocation={originPath ?? "/dashboard"}
+          />
         </form>
       </Inputs.Form>
       <FAQFooter />
