+++ src/services/ui/src/features/submission/waiver/contracting/contracting-1915-b-waiver-amendment.tsx
@@ -270,7 +270,11 @@ export const Contracting1915BWaiverAmendmentPage = () => {
             />
           </SectionCard>
           <Content.PreSubmissionMessage />
-          <SubmitAndCancelBtnSection />
+          <SubmitAndCancelBtnSection
+            showAlert
+            loadingSpinner
+            cancelNavigationLocation={originPath ?? "/dashboard"}
+          />
         </form>
       </Inputs.Form>
       <FAQFooter />
